+++ benchmark/src/main/java/benchmark/Throughput.java
@@ -1,0 +1,49 @@
+
+package benchmark;
+
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Fork;
+import org.openjdk.jmh.annotations.Measurement;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.annotations.Threads;
+import org.openjdk.jmh.annotations.Warmup;
+
+import com.github.f4b6a3.ksuid.Ksuid;
+import com.github.f4b6a3.ksuid.KsuidCreator;
+
+@Fork(1)
+@Threads(1)
+@State(Scope.Thread)
+@Warmup(iterations = 3)
+@Measurement(iterations = 5)
+@BenchmarkMode(Mode.Throughput)
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+public class Throughput {
+
+	@Benchmark
+	public UUID UUID_randomUUID() {
+		return UUID.randomUUID();
+	}
+
+	@Benchmark
+	public String UUID_randomUUID_toString() {
+		return UUID.randomUUID().toString();
+	}
+
+	@Benchmark
+	public Ksuid KsuidCreator_getKsuid() {
+		return KsuidCreator.getKsuid();
+	}
+
+	@Benchmark
+	public String KsuidCreator_getKsuid_toString() {
+		return KsuidCreator.getKsuid().toString();
+	}
+}